Move DelegateConnection result entity out of 930
