+++ 400-rest/src/test/java/software/wings/scheduler/DelegateDisconnectDetectorIteratorTest.java
@@ -52,8 +52,8 @@ import io.harness.service.intfc.DelegateCache;
 import software.wings.WingsBaseTest;
 import software.wings.beans.AwsConfig;
 import software.wings.beans.DelegateConnection;
+import software.wings.beans.DelegateConnectionResult;
 import software.wings.beans.TaskType;
-import software.wings.delegatetasks.validation.core.DelegateConnectionResult;
 import software.wings.service.impl.DelegateConnectionDao;
 import software.wings.service.impl.DelegateObserver;
 import software.wings.service.impl.DelegateServiceImpl;
