+++ 400-rest/src/test/java/software/wings/scheduler/FailDelegateTaskIteratorTest.java
@@ -46,8 +46,8 @@ import io.harness.rule.Owner;
 
 import software.wings.WingsBaseTest;
 import software.wings.beans.Account;
+import software.wings.beans.DelegateConnectionResult;
 import software.wings.beans.TaskType;
-import software.wings.delegatetasks.validation.core.DelegateConnectionResult;
 import software.wings.service.impl.DelegateTaskServiceClassicImpl;
 import software.wings.service.intfc.AssignDelegateService;
