+++ 400-rest/src/test/java/software/wings/service/delegate/DelegateTaskProcessTest.java
@@ -114,6 +114,7 @@ import software.wings.app.PortalConfig;
 import software.wings.beans.Account;
 import software.wings.beans.AccountStatus;
 import software.wings.beans.AwsConfig;
+import software.wings.beans.DelegateConnectionResult;
 import software.wings.beans.Environment;
 import software.wings.beans.GcpConfig;
 import software.wings.beans.GitConfig;
@@ -126,7 +127,6 @@ import software.wings.beans.delegation.TerraformProvisionParameters;
 import software.wings.beans.settings.helm.HelmRepoConfigValidationTaskParams;
 import software.wings.beans.yaml.GitCommand.GitCommandType;
 import software.wings.core.managerConfiguration.ConfigurationController;
-import software.wings.delegatetasks.validation.core.DelegateConnectionResult;
 import software.wings.helpers.ext.k8s.request.K8sClusterConfig;
 import software.wings.helpers.ext.k8s.request.K8sInstanceSyncTaskParameters;
 import software.wings.helpers.ext.url.SubdomainUrlHelperIntfc;
