+++ 400-rest/src/test/java/software/wings/service/impl/AssignDelegateServiceImplTest.java
@@ -89,13 +89,13 @@ import io.harness.service.intfc.DelegateCache;
 import software.wings.WingsBaseTest;
 import software.wings.beans.AwsAmiInfrastructureMapping;
 import software.wings.beans.AwsConfig;
+import software.wings.beans.DelegateConnectionResult;
+import software.wings.beans.DelegateConnectionResult.DelegateConnectionResultBuilder;
+import software.wings.beans.DelegateConnectionResult.DelegateConnectionResultKeys;
 import software.wings.beans.Environment;
 import software.wings.beans.InfrastructureMapping;
 import software.wings.beans.InfrastructureMappingType;
 import software.wings.beans.TaskType;
-import software.wings.delegatetasks.validation.core.DelegateConnectionResult;
-import software.wings.delegatetasks.validation.core.DelegateConnectionResult.DelegateConnectionResultBuilder;
-import software.wings.delegatetasks.validation.core.DelegateConnectionResult.DelegateConnectionResultKeys;
 import software.wings.service.impl.aws.model.AwsIamListInstanceRolesRequest;
 import software.wings.service.impl.aws.model.AwsIamRequest;
 import software.wings.service.intfc.DelegateSelectionLogsService;
