+++ 420-delegate-service/src/main/java/software/wings/service/intfc/AssignDelegateService.java
@@ -14,7 +14,7 @@ import io.harness.beans.DelegateTask;
 import io.harness.delegate.beans.Delegate;
 import io.harness.delegate.task.TaskFailureReason;
 
-import software.wings.delegatetasks.validation.core.DelegateConnectionResult;
+import software.wings.beans.DelegateConnectionResult;
 
 import java.util.List;
